feat:update fd optimizer branch for installable refactor
